Move integration test functions to a shared location
